Move funding reducer
